CRUD de anunciantes finalizado e implementado endpoints para crud de anúncios, assim como, ativar ou desativar um anúncio

+++ src/v1/DAO/AbstractDAO.php
@@ -12,28 +12,44 @@ abstract class AbstractDAO {
 			$this->entityManager = $entityManager;
 	}
 	
-	public function insert($obj){
+	public function inserir($obj){
 		$this->entityManager->persist($obj);
 		$this->entityManager->flush();
 	}
 
-	public function update($obj){
+	public function atualizar($obj){
 		$this->entityManager->merge($obj);
 		$this->entityManager->flush();
 	}
 
-	public function delete($obj){
-		$this->entityManager->remove($obj);
-		$this->entityManager->flush();
+	public function deletar($id){
+		$obj = $this->entityManager ->find($this->entityPath, $id);
+
+		if(!is_null(obj)){
+			$this->entityManager->remove($obj);
+			$this->entityManager->flush();
+		}
 	}
 
-	public function findById($id){
+	public function buscarPorId($id){
+		$obj = $this->entityManager ->find($this->entityPath, $id);
+
+		return is_null($obj) ? [] : $obj;
+	}
+
+	public function visualizarPorId($id){
 		$obj = $this->entityManager ->find($this->entityPath, $id);
 
 		return is_null($obj) ? [] : $obj->toArray();
 	}
 
-	public function findAll(){
+	public function identifica($id){
+		$obj = $this->entityManager ->find($this->entityPath, $id);
+
+		return !is_null($obj);
+	}
+
+	public function buscarTodos(){
 		$collection = $this->entityManager->getRepository($this->entityPath)->findAll();
 
 		$data = array();
@@ -42,5 +58,28 @@ abstract class AbstractDAO {
 		}
 
 		return $data;
+	}
+
+	public function buscarPor($array){
+		$collection = $this->entityManager->getRepository($this->entityPath)->findBy($array);
+
+		$data = array();
+		foreach($collection as $obj) {
+			$data [] = $obj->toArray();
+		}
+
+		return $data;
+	}
+
+	public function buscarUmPor($array){
+		$obj = $this->entityManager->getRepository($this->entityPath)->findOneBy($array);
+
+		return is_null($obj) ? [] : $obj;
+	}
+
+	public function visualizarUmPor($array){
+		$obj = $this->entityManager->getRepository($this->entityPath)->findOneBy($array);
+
+		return is_null($obj) ? [] : $obj->toArray();
 	}
 }
